解决warning

+++ EventLoop.cc
@@ -94,7 +94,7 @@ void EventLoop::handleRead() {
   uint64_t one = 1;
   ssize_t n = read(wakeupFd_, &one, sizeof one);
   if (n != sizeof one) {
-    LOG_ERROR("EventLoop::handleRead() reads %d bytes instead of 8", n);
+    LOG_ERROR("EventLoop::handleRead() reads %d bytes instead of 8", static_cast<int>(n));
   }
 }
